Cast the array index to int
git-svn-id: https://svn.parrot.org/parrot/trunk@20011 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/unicode.h
@@ -78,7 +78,7 @@ typedef unsigned long utf32_t;
 
 extern const char Parrot_utf8skip[256];
 
-#define UTF8SKIP(s) Parrot_utf8skip[*(s)]
+#define UTF8SKIP(s) Parrot_utf8skip[(int)*(s)]
 
 #define UTF8_MAXLEN 4
 #define UTF16_MAXLEN 4
